# RatingServices/src/main/resources/api-docs/openapi.yaml
openapi: 3.0.1
info:
  title: Rating API
  description: API for managing product ratings
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
  /ratings/{productId}:
    get:
      tags:
        - rating
      summary: Get all ratings for a specific product
      operationId: getAllRatings
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: List of product ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ratings'
  /ratings/rating/{id}:
    get:
      tags:
        - rating
      summary: Get a specific rating
      operationId: getRating
      parameters:
        - name: id
          in: path
          description: Unique identifier of the rating
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Rating details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ratings'
        "404":
          description: Rating not found
  /ratings/:
    post:
      tags:
        - rating
      summary: Create a new rating
      operationId: createRating
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ratings'
        required: true
      responses:
        "201":
          description: Rating successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ratings'
  /ratings/{id}:
    put:
      tags:
        - rating
      summary: Update a specific rating
      operationId: updateRating
      parameters:
        - name: id
          in: path
          description: Unique identifier of the rating
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ratings'
        required: true
      responses:
        "200":
          description: Rating successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ratings'
        "400":
          description: "Bad request (e.g., rating IDs do not match)"
        "404":
          description: Rating not found
  /ratings/{productId}/average:
    get:
      tags:
        - rating
      summary: Get average rating for a specific product
      operationId: getAverageRating
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: Average rating of the product
          content:
            application/json:
              schema:
                type: number
                format: double
        "404":
          description: Product not found
  /ratings/{deleteId}:
    delete:
      tags:
        - rating
      summary: Delete a specific rating
      operationId: deleteRating
      parameters:
        - name: deleteId
          in: path
          description: Unique identifier of the rating
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Rating successfully deleted
        "404":
          description: Rating not found
components:
  schemas:
    Ratings:
      required:
        - productId
        - score
      type: object
      properties:
        productId:
          type: string
          description: Unique identifier of the product
        score:
          type: integer
          description: Score of the product
          format: int32
        comment:
          type: string
          description: Optional comment for the rating